Extract language detection into own js module
Move setting the desired options into our own language detector.
